Implementacion de Value Object y Persistencia de Datos

+++ 4_CREACION_TRANSFERENCIA/transfer-service/src/main/java/com/digitalwallet/transferservice/transfer/domain/TransferRequest.java
@@ -2,11 +2,12 @@ package com.digitalwallet.transferservice.transfer.domain;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class TransferRequest {
-
     private String originPhoneNumber;
     private String destinationPhoneNumber;
-    private double amount;
+    private BigDecimal amount;
     private String currency;
 }
